requirement.txt
Flask==2.0.1
Pillow==8.3.1
tensorflow==2.5.0
